2.4 (#40)
* framework for generic sync values

* docs & byte buf string helpers

* rewrite scrolling

* alt+ctrl+l

* scroll fixes & improvements

* Refactor screen opening backend (#39)

* extract color objects to its own ColorShade class

* clean up texture code

* fix crash & dont animate if a screen is already open

* make open methods static

* adaptable textures can have different border thickness on each side

* minor changes

* use gt5 cover button textures as default button textures

* clean up

* fix some ide warnings

* even more clean up

+++ src/main/java/com/cleanroommc/modularui/drawable/TextRenderer.java
@@ -76,7 +76,7 @@ public class TextRenderer {
         for (Line measuredLine : measuredLines) {
             int x0 = getStartX(measuredLine.width);
             maxW = Math.max(draw(measuredLine.text, x0, y0), maxW);
-            y0 += getFontHeight();
+            y0 += (int) getFontHeight();
         }
         this.lastWidth = this.maxWidth > 0 ? Math.min(maxW, this.maxWidth) : maxW;
         this.lastHeight = measuredLines.size() * getFontHeight();
@@ -125,7 +125,7 @@ public class TextRenderer {
             drawMeasuredLines(Collections.singletonList(line));
             return;
         }
-        scroll = scroll % (int) (line.width + 1) ;
+        scroll = scroll % (int) (line.width + 1);
         String drawString = line.getText();//getFontRenderer().trimStringToWidth(line.getText(), (int) (this.maxWidth + scroll));
         Area.SHARED.set(this.x, Integer.MIN_VALUE, this.x + (int) this.maxWidth, Integer.MAX_VALUE);
         Stencil.apply(Area.SHARED, context);
